# docs/swagger.yaml
basePath: /api/v1
definitions:
  helpers.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  models.InsertComment:
    properties:
      message:
        example: Nice Info!
        type: string
      photo_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
...
  models.UpdateComment:
    properties:
      message:
        example: I Wanna Hire You. Please, check your DM!
        type: string
      photo_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: iqbaludinm14@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample service for social-media.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MyGram API
  version: "1.0"
paths:
...
  /comments:
    get:
      consumes:
      - application/json
      description: Get list of comments based on query params of photo id
      parameters:
      - description: Photo ID
        in: query
        name: photo_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - ApiKeyAuth: []
      summary: Get All Comment by Photo Id
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: Create a comment and store to database
      parameters:
      - description: Photo ID
        in: query
        name: photo_id
        required: true
        type: integer
      - description: Field for insert comment
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/models.InsertComment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - ApiKeyAuth: []
      summary: Create Comment
      tags:
      - Comment
  /comments/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete Comment
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Comment
      tags:
      - Comment
    get:
      consumes:
      - application/json
      description: Get a comment by id
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Comment By Id
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: Update a comment
      parameters:
      - description: Comment Id
        in: path
        name: commentId
        required: true
        type: string
      - description: Update a Comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.UpdateComment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Comment
      tags:
      - Comment
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"